Replaced URL to 'feedafever.com/anon/' with `rel="noreferrer"`
Amends 5bb01243662e1f7ce81210262a7d7cda7c2691dc

Since feedafever.com is offline, the `anonymize` preference is unusable since it would attempt to make a request to 'feedafever.com/anon/?go=*'.

This commit replaces the obsolete intermediary URL with the `noreferrer` keyword for the `rel` attribute which instructs the browser, when navigating to the target resource, to omit the `Referer` HTTP header.

Furthermore, regardless of the `anonymize` preference, Fever will instruct the browser to navigate to the target resource without granting the new browsing context access to the document that opened it via the `noopener` keyword which the `noreferrer` also includes.